backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3

# Database dependencies
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1